Dry deposition, emission, and vertical diffusion routines where updated to use a common shared data module, ASX_DATA_MOD. This reduces a number of redundant memory allocations, reads from file, and calculations of environmental variables. FORTRAN derived data types are used in the ASX_DATA_MOD.
Two errors were found in BEIS 3 and corrected. (1) The call to CLNEW_SUB depended on either the solar zenith angle or PAR. This resulted in divide by zero errors and an overestimate of PAR under low solar zenith angles due to the application of the Campbell and Norman (1998) canopy light environment algorithms outside of their valid range. (2) The Cl variable from Guenther et al 1999 was missing its dependence on the canopy LAI.

The size-correction factors for sea salt emissions from Zhang et al 2005 where updated to Equation 3 of Lewis & Schwartz 2006 and Zhang et al 2006 equation 2.

M3dry and MOSAIC routines were extensively rewritten. An error in the estimation of agricultural LAI was corrected.

The VDIFF_MET module was renamed and expanded to the ASX_DATA_MOD due to now being applied in more areas of the code than vertical diffusion.

The identification and use of land use categories between CMAQ and WRF-CMAQ were harmonized allowing for the simplification of code in areas where land use information is used.

	modified:   ../../aero/aero5/aero_depv.F
	modified:   ../../aero/aero6/AERO_EMIS.F
	modified:   ../../biog/beis3/beis3.F
	modified:   ../../biog/beis3/czangle.F
	modified:   ../../biog/beis3/hrno.F
	modified:   ../../biog/beis3/tmpbeis.F
	modified:   ../../cloud/cloud_acm_ae6/rescld.F
	modified:   ../../depv/m3dry/ABFLUX_MOD.F
	modified:   ../../depv/m3dry/BIDI_MOD.F
	modified:   ../../depv/m3dry/DEPVVARS.F
	modified:   ../../depv/m3dry/DEPV_DEFN.F
	modified:   ../../depv/m3dry/LSM_MOD.F
	modified:   ../../depv/m3dry/MOSAIC_MOD.F
	modified:   ../../depv/m3dry/m3dry.F
	modified:   ../../emis/emis/EMIS_DEFN.F
	modified:   ../../emis/emis/SSEMIS.F
	modified:   ../../phot/phot_inline/phot.F
	new file:   ../../vdiff/acm2/ASX_DATA_MOD.F
	deleted:    ../../vdiff/acm2/VDIFF_MET.F
	modified:   ../../vdiff/acm2/aero_sedv.F
	modified:   ../../vdiff/acm2/eddyx.F
	modified:   ../../vdiff/acm2/sedi.F
	modified:   ../../vdiff/acm2/vdiffacmx.F
	modified:   ../../vdiff/acm2/vdiffproc.F
	new file:   ../../vdiff/acm2_mp/ASX_DATA_MOD.F
	deleted:    ../../vdiff/acm2_mp/VDIFF_MET.F
	modified:   ../../vdiff/acm2_mp/vdiffproc.F