modified:   biog/beis3/beis3.F
Revised the photosynthetic photon flux density function on light dependent VOC emissions to include the dependency of light attenuation through the vegetation canopy and expand it to the two layer sun and shade leaf parameterization in BEIS 3.6. Updated to Niinemets et al. 2010 doi:10.1029/2010JG001436 Big leaf model which expands the Guenther et al. 1993 doi:10.1029/93JD00527 parameterization for in-canopy gradients. This results in lower light dependent VOC emissions in the morning and evening transition and approximately a 40% increase in these emissions during midday resulting in a slight increase, ~10%, in the total daily emissions.

	modified:   vdiff/acm2/ASX_DATA_MOD.F

Updated the diffusivities, LaBas molar volumes, mesophyll resistance, and cuticular resistance apparent reactivity for organic nitrates in CB05e51 and SAPRAC.