Coarse Mode Hplus Patch
Sarwar found that coarse mode H+ concentrations were 0 in runs with the code from the master branch. This problem can be traced to an issue at line 1557 in aero_subs.F where the local HPLUS concentrations were being initialized to 0.0 after the coarse mode H+ concentrations were calculated. The simple fix of course is to only initialize the aitken and accumulation modes. This fix should mostly affect heterogeneous chemistry processes, although the magnitude of that affect is not certain and potentially somewhat small.