Changed caer( im ) to caer ( i )
Note that this change was only detected when all science processes were commented out in sciproc.F except for AERO so the model encountered the "threshold not exceeded" case. The update was made because caer( i ) was used after the "End Do" statement for the loop over i. As a result, the model crashed from a segmentation fault.