Overview:
This commit fixes an error in the diagnostic biogenic volatile organic compound emissions file using the WRF-CMAQ two way coupled model. There is currently an error in the output time step of this file. When running in the two way coupled model somewhere between 1 and 24 time steps are written to file for a one-day simulation with hourly output. This correction ensures that the number of output time steps is consistent with the other CMAQ output files.

Details:

This was accomplished by modifying the WSTEP variable in the tmpbeis.F subroutine to be calculated in the same way as other subroutines that write diagnostic output. This was tested for both online and offline simulations.

Impact:

The biogenic VOC emission file now has the requested number of output time steps when run inline. There is no impact on the model results.

	modified:   CCTM/src/biog/beis3/tmpbeis.F