Overview
The main goal of this request is to correct a systematic underestimate of wet deposition of all species when running the WRF-CMAQ coupled model and enable the output of wet deposition fields on a user defined time step. This bug fix only changes the results in the WRF-CMAQ coupled model.

Details

The wet deposition bug in the WRF-CMAQ coupled model was related to the writing of the output time step in cldproc_acm.F and setting the accumulated hour wet deposition for a grid cell to zero if the precipitation rate or cloud fraction fell below their respective threshold values in cldproc_acm.F. The logic governing the write time in cldproc_acm.F was revised by copying the logic in the aerosol and vdiff code for diagnostic outputs. The portion of the code in cldproc_acm.F where the accumulated wet deposition array was set to zero was removed because this array is re-initialized to zero at each time the output file is written. This was a redundant operation that was hardcoded for hourly meteorological inputs.  opwdep.F was also modified to allow the WDEP2 diagnostic file to support a user defined output time step.

Impact

WRF-CMAQ coupled wet deposition results running with a 5 minute communication between WRF and CMAQ were increased by approximately 30% for all species. This matches offline simulation much better with approximately a ±5% difference between online and offline deposition fields for the simulation day tested. All WRF-CMAQ simulations to date are impacted by this bug.

	modified:   CCTM/src/cloud/acm_ae6/cldproc_acm.F
	modified:   CCTM/src/cloud/acm_ae6/convcld_acm.F
	modified:   CCTM/src/cloud/acm_ae6/opwdep.F
	modified:   CCTM/src/cloud/acm_ae6_mp/cldproc_acm.F
	modified:   CCTM/src/cloud/acm_ae6_mp/convcld_acm.F