Bugfix to writesite utility 	modified:   POST/writesite/src/writesite.F
Bugfix by Christian Hogrefe:
The previous version of writesite (released with CMAQv5.2) did not support polar stereographic projections except when SITE_FILE was set to “ALL” (or not set at all, since the default behavior seems to be to internally set SITE_FILE to ALL if not set to a different value in the script).

This revised version added polar stereographic projection support to the “ToProj” subroutine in writesite.F which is called by module_site.F if SITE_FILE is set to a value different than ALL (and the corresponding SITE_FILE contains valid lon/lat entries).

In addition, the error messages at the end of the SetProj, ToProj, and ToLL subroutines in writesite.F were updated to say “Unsupported map projection” instead of “Map projection setup error”