Added functionality for region-dependent scaling
The improvements here add subroutines for reading, processing, and
applying gridded masks identifying particular subsets of the model
domain. The logic for implementing the emissions scaling rules had
to be updated and expanded to accomodate the spatially-dependent
scaling feature. In order to accomplish this, each operation ("a",
"m" and "o") for a combinatino of instructino and source are stored
as a function of the region the operation is applied to. Then the
actual scale factors are combined at the same time that they are
applied to the emission rates, as a function of species, source,
and now, column, row and layer.

	modified:   CCTM/src/MECHS/cb6r3_ae6_aq/EmissCtrl.nml
	modified:   CCTM/src/emis/emis/EMIS_DEFN.F
	modified:   CCTM/src/emis/emis/EMIS_VARS.F
	modified:   CCTM/src/vdiff/acm2/vdiffproc.F
	modified:   bldit_project.csh