modified:   tmpbeis.F
Importance: High (bug fix)

Why would you need it?
This bug fix is needed to output the correct biogenic emissions when running BEIS inline.

Impact on Results?
This only impact the biogenic emissions diagnostic output. The CMAQ model results are unchanged.

The biogenic emissions diagnostic file was inadvertently broken when it was updated to output emissions with the WRF/CMAQ coupled model. This bug fix corrects the introduced bug. Additionally, emissions were time weighted in the diagnostic file and summed when passed to the chemical transport model. The code has been changed to output the summed emissions in the same way as the emissions passed to the CTM and then converted units specified in the output time step.