Improved documentation in LSM_MOD, BIDI_MOD, and  NH3_BIDI_MOD. 	modified:   CCTM/src/depv/stage/BIDI_MOD.F         modified:   CCTM/src/depv/stage/NH3_BIDI_MOD.F 	modified:   CCTM/src/depv/stage/LSM_MOD.F
Implemented the Massad et al. 2010 land use specific cuticular resistance for NH3 when using bidirectional exchange. Vectorized the tiled aerosol deposition code to be consistent with the m3dry implementation.
	modified:   CCTM/src/depv/stage/STAGE_MOD.F

Added the missing impact of heterogeneous HONO on NO2 deposition and deposition velocities for tiled outputs. This is now consistent with the grid cell fluxes.
	modified:   CCTM/src/vdiff/acm2_stage/vdiffacmx.F