Updates to Chapter 6 - dry deposition
The first paragraph of section 6.8.2 had text that was really more about m3dry.  I moved that text to the m3dry section and made a more general introduction to the section.  There is still some general text needed on mercury bidirectional exchange in the top section and it would be great to get input from @jessebash on that.