Remove CCTM_VDIFF_DIAG from Appendix A of UG
This file cannot be created with the current code and runscript so removed the description from the User's Guide Appendix A (already removed from Chapter 7). If this diagnostic file is of interest down the road someone can debug the current issue and then we can add it back to documentation.