Fixed typo in User's Guide Ch 9 (Process Analysis)
In the CMAQ User's Guide section 9.6 "Example Applications": 
The last two bullets define variables for the loss of anthropogenic VOC (a previously-defined family) with O3 and with NO3. But the RHS of each equation does not match the species name:
•	IRR_OUTPUT aVOCwithO3 = LOSS[aVOC] AND [NO3];
•	IRR_OUTPUT aVOCwithNO3 = LOSS[aVOC] AND [O3];
It should be the other way around:
•	IRR_OUTPUT aVOCwithO3 = LOSS[aVOC] AND [O3];
•	IRR_OUTPUT aVOCwithNO3 = LOSS[aVOC] AND [NO3];
This typo is now corrected.  

A related set of changes is also needed in CCTM/src/MECHS/cb6r3_ae7_aq/pa_cb6r3_ae7_aq.ctl 
The updates to the control file will be included in the next CMAQ release.