Clarified the solution
Removed the statement that in CMAQv5.3, no changes to the species definition files are needed for the STAGE bi-directional option if downward deposition fluxes are desired. Changes are needed for both options, but the exact changes differ due to the different naming conventions as described in the final two sentences of the solution.