One-line bug fix for emission control filename error catch
Hey Sergey, Christian stumbled into this error this morning. Basically, if you provide the wrong name for the emission control file, the code crashes and gives you no good reason. This variable just needs to be updated. We made the fix for a pull request that went to the research branch, but that didn't make it to master and Christian was using master. It cost him some time and would be great to get out there to the user community. Do you mind accepting it with this pull request?