sitecmp_dailyo3 CASTNET QA_FLAG processing bug fix
This commit fixes a bug in sitecmp_dailyo3 that caused
CASTNET observations with QA flags indicating bad
measurements to still be considered in computing the
daily ozone metrics. The bug was caused by applying the
rmQuots function to the string holding the ozone
concentration rather than the string holding the
QA flag. Therefore, when comparing the QA flag string
against the master string with all QA flags to be
tested against, no match was found if the QA flag
in the observation file was in quotes and the ozone
value was passed through to the subsequent computations.
This only affected hours 2-24 for any given day, the
coding for the first hour was correct.

	modified:   utilities.F

(cherry picked from commit a78c58559828caf14c3493557ad08ddba019b071
to prepare an equivalent of PR#732 as PR to the research branch)