addressed assignment statement flagged by gcc debug
When testing the previous commits with gcc debug, the run crashed
with this error message:
Fortran runtime error: Array bound mismatch for dimension 1 of array 'pts' (50/20)

Error termination. Backtrace:
        at /work/MOD3EVAL/css/git_builds/test_hr2day_updates_cmaqv54_research/POST/hr2day/scripts/BLD_hr2day_v532_gcc6.1_debug/getTZ.F:226

While this doesn't appear to be an subscript out-of-bound situation since
the number of elements in the RHS expression is less than the number of elements
allocated for the LHS expression, the following change appears to address the issue:

Original code:

       pts(i)%x(:,count(i)) = lon(1:lines(nlines)%npts)
       pts(i)%y(:,count(i)) = lat(1:lines(nlines)%npts)

Revised code:

       pts(i)%x(1:lines(nlines)%npts,count(i)) = lon(1:lines(nlines)%npts)
       pts(i)%y(1:lines(nlines)%npts,count(i)) = lat(1:lines(nlines)%npts)

Note that the original code is the one as it currently exists in the
CMAQ_Dev and CMAQ repo, so it likely wasn't tested previously with running
a post-processing case in gcc debug mode.

	../../../bldit_project_scratch.csh