Update PULL_REQUEST_TEMPLATE.md
Minor change to PR template to add lines for Primary and Secondary Contacts to the Release Notes Section.  The Secondary Contact should be used when the Primary Contact is in a temporary position (e.g., student or postdoc).  This can also be used to list multiple developers as POCs when appropriate.