abort if DDM and STAGE are specified
DDM does not currently support the STAGE deposition model. If the user attempts to build with both options, the code will not compile. This can be confusing and time-wasting. This PR puts in a check and aborts with a message if the user selects these incompatible options. Also, I noticed that our script usage instructions at the top are wrong and fixed them.